大师行为

+++ AssetReader/Brotli/Decode.cs
@@ -22,7 +22,7 @@ namespace Org.Brotli.Dec {
 
         private const int HuffmanTableBits = 8;
 
-        private const int HuffmanTableMask = unchecked((int)(0xFF));
+        private const int HuffmanTableMask = unchecked(0xFF);
 
         private const int CodeLengthCodes = 18;
 
@@ -35,9 +35,9 @@ namespace Org.Brotli.Dec {
         private static readonly int[] DistanceShortCodeValueOffset = new int[] { 0, 0, 0, 0, -1, 1, -2, 2, -3, 3, -1, 1, -2, 2, -3, 3 };
 
         /// <summary>Static Huffman code for the code length code lengths.</summary>
-        private static readonly int[] FixedTable = new int[] { unchecked((int)(0x020000)), unchecked((int)(0x020004)), unchecked((int)(0x020003)), unchecked((int)(0x030002)), unchecked((int)(0x020000)), unchecked((int)(0x020004)), unchecked((int)(0x020003
-            )), unchecked((int)(0x040001)), unchecked((int)(0x020000)), unchecked((int)(0x020004)), unchecked((int)(0x020003)), unchecked((int)(0x030002)), unchecked((int)(0x020000)), unchecked((int)(0x020004)), unchecked((int)(0x020003)), unchecked((int
-            )(0x040005)) };
+        private static readonly int[] FixedTable = new int[] { unchecked(0x020000), unchecked(0x020004), unchecked(0x020003), unchecked(0x030002), unchecked(0x020000), unchecked(0x020004), unchecked(0x020003
+            ), unchecked(0x040001), unchecked(0x020000), unchecked(0x020004), unchecked(0x020003), unchecked(0x030002), unchecked(0x020000), unchecked(0x020004), unchecked(0x020003), unchecked(
+            0x040005) };
 
         /// <summary>Decodes a number in the range [0..255], by reading 1 - 11 bits.</summary>
         private static int DecodeVarLenUnsignedByte(BitReader br) {
@@ -97,7 +97,7 @@ namespace Org.Brotli.Dec {
             var val = (int)((long)(((ulong)br.accumulator) >> br.bitOffset));
             offset += val & HuffmanTableMask;
             var bits = table[offset] >> 16;
-            var sym = table[offset] & unchecked((int)(0xFFFF));
+            var sym = table[offset] & unchecked(0xFFFF);
             if (bits <= HuffmanTableBits) {
                 br.bitOffset += bits;
                 return sym;
@@ -106,7 +106,7 @@ namespace Org.Brotli.Dec {
             var mask = (1 << bits) - 1;
             offset += (int)(((uint)(val & mask)) >> HuffmanTableBits);
             br.bitOffset += ((table[offset] >> 16) + HuffmanTableBits);
-            return table[offset] & unchecked((int)(0xFFFF));
+            return table[offset] & unchecked(0xFFFF);
         }
 
         private static int ReadBlockLength(int[] table, int offset, BitReader br) {
@@ -139,7 +139,7 @@ namespace Org.Brotli.Dec {
                 mtf[i] = i;
             }
             for (var i = 0; i < vLen; i++) {
-                var index = v[i] & unchecked((int)(0xFF));
+                var index = v[i] & unchecked(0xFF);
                 v[i] = unchecked((byte)mtf[index]);
                 if (index != 0) {
                     MoveToFront(mtf, index);
@@ -160,7 +160,7 @@ namespace Org.Brotli.Dec {
                 BitReader.FillBitWindow(br);
                 var p = (int)(((long)(((ulong)br.accumulator) >> br.bitOffset))) & 31;
                 br.bitOffset += table[p] >> 16;
-                var codeLen = table[p] & unchecked((int)(0xFFFF));
+                var codeLen = table[p] & unchecked(0xFFFF);
                 if (codeLen < CodeLengthRepeatCode) {
                     repeat = 0;
                     codeLengths[symbol++] = codeLen;
@@ -269,7 +269,7 @@ namespace Org.Brotli.Dec {
                     var p = (int)((long)(((ulong)br.accumulator) >> br.bitOffset)) & 15;
                     // TODO: Demultiplex FIXED_TABLE.
                     br.bitOffset += FixedTable[p] >> 16;
-                    var v = FixedTable[p] & unchecked((int)(0xFFFF));
+                    var v = FixedTable[p] & unchecked(0xFFFF);
                     codeLengthCodeLengths[codeLenIdx] = v;
                     if (v != 0) {
                         space -= (32 >> v);
@@ -354,7 +354,7 @@ namespace Org.Brotli.Dec {
             DecodeBlockTypeAndLength(state, 0);
             var literalBlockType = state.blockTypeRb[1];
             state.contextMapSlice = literalBlockType << LiteralContextBits;
-            state.literalTreeIndex = state.contextMap[state.contextMapSlice] & unchecked((int)(0xFF));
+            state.literalTreeIndex = state.contextMap[state.contextMapSlice] & unchecked(0xFF);
             state.literalTree = state.hGroup0.trees[state.literalTreeIndex];
             int contextMode = state.contextModes[literalBlockType];
             state.contextLookupOffset1 = Context.LookupOffsets[contextMode];
@@ -373,7 +373,7 @@ namespace Org.Brotli.Dec {
 
         private static void MaybeReallocateRingBuffer(State state) {
             var newSize = state.maxRingBufferSize;
-            if ((long)newSize > state.expectedTotalSize) {
+            if (newSize > state.expectedTotalSize) {
                 /* TODO: Handle 2GB+ cases more gracefully. */
                 var minimalNewSize = (int)state.expectedTotalSize + state.customDictionary.Length;
                 while ((newSize >> 1) > minimalNewSize) {
@@ -624,14 +624,14 @@ namespace Org.Brotli.Dec {
                                 }
                             }
                         } else {
-                            var prevByte1 = ringBuffer[(state.pos - 1) & ringBufferMask] & unchecked((int)(0xFF));
-                            var prevByte2 = ringBuffer[(state.pos - 2) & ringBufferMask] & unchecked((int)(0xFF));
+                            var prevByte1 = ringBuffer[(state.pos - 1) & ringBufferMask] & unchecked(0xFF);
+                            var prevByte2 = ringBuffer[(state.pos - 2) & ringBufferMask] & unchecked(0xFF);
                             while (state.j < state.insertLength) {
                                 BitReader.ReadMoreInput(br);
                                 if (state.blockLength[0] == 0) {
                                     DecodeLiteralBlockSwitch(state);
                                 }
-                                var literalTreeIndex = state.contextMap[state.contextMapSlice + (Context.Lookup[state.contextLookupOffset1 + prevByte1] | Context.Lookup[state.contextLookupOffset2 + prevByte2])] & unchecked((int)(0xFF));
+                                var literalTreeIndex = state.contextMap[state.contextMapSlice + (Context.Lookup[state.contextLookupOffset1 + prevByte1] | Context.Lookup[state.contextLookupOffset2 + prevByte2])] & unchecked(0xFF);
                                 state.blockLength[0]--;
                                 prevByte2 = prevByte1;
                                 BitReader.FillBitWindow(br);
@@ -662,7 +662,7 @@ namespace Org.Brotli.Dec {
                             }
                             state.blockLength[2]--;
                             BitReader.FillBitWindow(br);
-                            state.distanceCode = ReadSymbol(state.hGroup2.codes, state.hGroup2.trees[state.distContextMap[state.distContextMapSlice + (state.copyLength > 4 ? 3 : state.copyLength - 2)] & unchecked((int)(0xFF))], br);
+                            state.distanceCode = ReadSymbol(state.hGroup2.codes, state.hGroup2.trees[state.distContextMap[state.distContextMapSlice + (state.copyLength > 4 ? 3 : state.copyLength - 2)] & unchecked(0xFF)], br);
                             if (state.distanceCode >= state.numDirectDistanceCodes) {
                                 state.distanceCode -= state.numDirectDistanceCodes;
                                 var postfix = state.distanceCode & state.distancePostfixMask;
